# openapi.yml
openapi: 3.0.3
info:
  title: ACMS - Apartment Committee Management System
  description: |
    Multi-tenant SaaS platform for apartment/housing societies.
    Five role types: site_admin, apartment_admin, committee_head, committee_member, resident.
    Tenant isolation via `apartmentId` — all non-site-admin queries are scoped.
  version: 1.0.0
  contact:
    name: ACMS Team
servers:
  - url: /api
    description: API base path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VisitorLog:
      type: object
      properties:
        _id: { type: string }
        apartmentId: { type: string }
        visitorName: { type: string }
        purpose: { type: string }
        unitVisited: { type: string }
        loggedBy: { type: string, description: 'User ID or populated' }
        checkIn: { type: string, format: date-time }
        checkOut: { type: string, format: date-time, nullable: true }
    CreateVisitorLogRequest:
      type: object
      required: [visitorName, purpose, unitVisited]
      properties:
        visitorName: { type: string }
        purpose: { type: string }
        unitVisited: { type: string, example: A-101 }

    # ────────── Notice ──────────
...
paths:
...
  /visitors:
    get:
      tags: [Visitor Logs]
      summary: List visitor logs (scoped by role)
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/VisitorLog'
    post:
      tags: [Visitor Logs]
      summary: Log a visitor entry (committee member/head)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVisitorLogRequest'
      responses:
        '201': { description: Visitor logged }
  /visitors/{id}/checkout:
    put:
      tags: [Visitor Logs]
      summary: Check out a visitor
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Visitor checked out }
        '404': { description: Log not found or already checked out }

  # ══════════════════════════════════════════════════
  # NOTICES
  # ══════════════════════════════════════════════════
...